ui/requirements.txt
streamlit
requests
Pillow
streamlit-lottie